requirements.txt
absl-py==0.1.12
appnope==0.1.0
argon2-cffi==20.1.0
astor==0.6.2
async-generator==1.10
attrs==20.3.0
bleach==1.5.0
cffi==1.14.4
cycler==0.10.0
decorator==4.2.1
defusedxml==0.6.0
entrypoints==0.3
gast==0.2.0
graphviz==0.8.2
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
importlib-metadata==3.1.0
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.11.1
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras==2.1.5
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.1.1
matplotlib==2.2.2
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.14.2
pandocfilters==1.4.3
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
prometheus-client==0.9.0
prompt-toolkit==1.0.18
protobuf==3.5.2.post1
ptyprocess==0.5.2
pycparser==2.20
pydot==1.2.4
pydot-ng==1.0.0
Pygments==2.7.2
pyparsing==2.2.0
pyrsistent==0.17.3
python-chess==0.31.4
python-dateutil==2.7.1
pytz==2018.3
PyYAML==3.12
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
scipy==1.0.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tornado==5.0.1
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.5.1
zipp==3.4.0
